requirements.txt
python-telegram-bot[job-queue]==20.8
gspread==5.12.2
google-auth-oauthlib==1.2.0
google-generativeai==0.7.0
matplotlib==3.8.4
requests==2.32.3
firebase-admin==6.5.0
Flask==3.0.3
google-auth
telegram